# backend/docs/swagger.yaml
definitions:
...
  keys.CreateKeyRequest:
    properties:
      iv:
        type: string
      label:
        type: string
      passphrase:
        type: string
      value:
        type: string
    required:
    - iv
    - label
    - value
    type: object
  keys.KeyResponse:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      iv:
        type: string
      label:
        type: string
      passphrase:
        type: string
      updatedAt:
        type: string
      value:
        type: string
    type: object
  keys.UpdateKeyRequest:
    properties:
      iv:
        type: string
      label:
        type: string
      passphrase:
        type: string
      value:
        type: string
    type: object
  models.Error:
    properties:
      errors:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
      statusCode:
        type: integer
    type: object
...
info:
  contact: {}
  description: API server for FlySsh
  title: FlySsh API
  version: "1.0"
paths:
...
  /keys:
    get:
      description: Return all SSH keys belonging to the authenticated user.
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/keys.KeyResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
      summary: List all SSH keys
      tags:
      - keys
    post:
      consumes:
      - application/json
      description: Store a client-encrypted SSH key. The value and optional passphrase
        are opaque encrypted blobs.
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Key details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/keys.CreateKeyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/keys.KeyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Create a new SSH key
      tags:
      - keys
  /keys/{id}:
    delete:
      description: Delete an SSH key by ID. Associated hosts will have their keyId
        set to null.
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Key ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Error'
      summary: Delete an SSH key
      tags:
      - keys
    get:
      description: Return an SSH key by ID. Must belong to the authenticated user.
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Key ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/keys.KeyResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Error'
      summary: Get a single SSH key
      tags:
      - keys
    put:
      consumes:
      - application/json
      description: Update fields of an existing SSH key. All fields are optional.
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Key ID
        in: path
        name: id
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/keys.UpdateKeyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/keys.KeyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Error'
      summary: Update an SSH key
      tags:
      - keys
swagger: "2.0"